% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_overlap.R
\name{proportion_overlap}
\alias{proportion_overlap}
\title{Proportion overlap}
\usage{
proportion_overlap(base, comp, mn, mx)
}
\arguments{
\item{base, comp}{Vector of flow, velocity, or stage values for baseline and comparison scenarios}

\item{mn, mx}{Minimum and maximum values across all comparisons in the analysis}
}
\description{
Calculates proportion overlap in density distributions for two scenarios.
}
\examples{

}
