% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sac_sal.R
\docType{data}
\name{sac_sal}
\alias{sac_sal}
\title{Sacramento River salvage}
\format{A tibble with 749 rows and 32 columns}
\usage{
sac_sal
}
\description{
A dataset on number of fish salvaged at CVP and SWP based on releases of tagged fish in the North Delta.
}
\keyword{datasets}
