% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvage_model.R
\name{salvage_model}
\alias{salvage_model}
\title{Salvage model}
\usage{
salvage_model(run = "winter", facility = "both")
}
\arguments{
\item{run}{Chinook Salmon run: fall, late_fall, winter}

\item{facility}{Water export facility: CVP, SWP, both}
}
\description{
Zero-inflated model of salvage as function of Freeport flow, exports, and fish length.
}
