% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvage_newdata.R
\name{salvage_newdata}
\alias{salvage_newdata}
\title{Salvage new data}
\usage{
salvage_newdata(flow_df, annual_passage)
}
\arguments{
\item{flow_df}{Data frame with modeled flow input}

\item{annual_passage}{Arbitrary annual passage used salvage predictions}
}
\description{
Prepare flow_df data frame to be used for salvage predictions
}
