% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/winter_run_size.R
\docType{data}
\name{winter_run_size}
\alias{winter_run_size}
\title{Winter run size}
\format{A tibble with 9 rows and 2 columns
#' \describe{
  \item{Month}{numeric month}
  \item{Length}{fork length (mm)}
}}
\usage{
winter_run_size
}
\description{
Estimates of juvenile winter run fork lengths by month; used in salvage predictions.
}
\keyword{datasets}
