% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_score.R
\name{z_score}
\alias{z_score}
\title{Z-score}
\usage{
z_score(x, x_mean, x_sd)
}
\arguments{
\item{x}{Numeric vector}

\item{x_mean}{Mean of numeric vector (or mean of original data)}

\item{x_sd}{Standard deviation of numeric vector (or sd of original data)}
}
\description{
Calculate z-score
}
\examples{
z_score(1:10, mean(1:10), sd(1:10))
z_score(1:10, 20, 5)
}
