% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salvage_prediction.R
\name{salvage_prediction}
\alias{salvage_prediction}
\title{Salvage prediction}
\usage{
salvage_prediction(run, facility, newdata)
}
\arguments{
\item{run}{Chinook Salmon run: Fall, LateFall, Winter}

\item{facility}{Water export facility: CVP, SWP, both}

\item{newdata}{Data frame with input data needed for model predictions}
}
\description{
Predicted salvage based on salvage_newdata and salvage_model.
}
